requirements.txt
discord.py==2.4.0
pymysql==1.1.1
python-dotenv==1.0.1
rapidfuzz==3.9.6
python-dateutil==2.9.0.post0
